Handle multiple rules defined on one line.
Support multiple rules defined on a single line. Change global_macros to
work with the changes.

N.B.: If multiple rules are defined on one line, where one rule already
uses a global_macro, and the other could use the exact same global_macro,
the plugin will not suggest it, as the detected macro usage on that line
will "shadow" it.